add authorization to operations

+++ backend/controllers/goalController.js
@@ -1,13 +1,15 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel');
-const { request } = require('http');
+const User = require('../models/userModel');
 
 // @desc    get goals
 // @route   GET /api/goals
 // @access  private 
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({
+        user: req.user.id
+    });
 
     res.status(200).json(goals);
 });
@@ -23,6 +25,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
+        user: req.user.id
     })
 
     res.status(200).json(goal);
@@ -38,6 +41,20 @@ const updateGoal = asyncHandler(async (req, res) => {
         res.status(400);
 
         throw new Error('Goal not found');
+    }
+
+    const user = await User.findById(req.user.id);
+
+    // check for user
+    if(!user){
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // make sure logged in user matches goal user 
+    if(goal.user.toString() !== user.id){
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -57,6 +74,20 @@ const deleteGoal = asyncHandler(async (req, res) => {
         res.status(400);
 
         throw new Error('Goal not found');
+    }
+
+    const user = await User.findById(req.user.id);
+
+    // check for user
+    if(!user){
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // make sure logged in user matches goal user 
+    if(goal.user.toString() !== user.id){
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     await goal.deleteOne();
